requirements.txt
fastapi
uvicorn
pydantic
presidio-analyzer
presidio-anonymizer
spacy
python-docx
pypdf
python-multipart
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl